+++ modal/partial_function.py
@@ -89,6 +89,9 @@ class _PartialFunction(typing.Generic[P, ReturnType, OriginalReturnType]):
         self.force_build = force_build
         self.build_timeout = build_timeout
 
+    def _get_raw_f(self) -> Callable[P, ReturnType]:
+        return self.raw_f
+
     def __get__(self, obj, objtype=None) -> _Function[P, ReturnType, OriginalReturnType]:
         k = self.raw_f.__name__
         if obj:  # accessing the method on an instance of a class, e.g. `MyClass().fun``
@@ -551,6 +554,14 @@ def _build(
     """
     if _warn_parentheses_missing is not None:
         raise InvalidError("Positional arguments are not allowed. Did you forget parentheses? Suggestion: `@build()`.")
+
+    deprecation_warning(
+        (2025, 1, 15),
+        "The `@modal.build` decorator is deprecated and will be removed in a future release."
+        "\n\nWe now recommend storing large assets (such as model weights) using a `modal.Volume`"
+        " instead of writing them directly into the `modal.Image` filesystem."
+        "\n\nSee https://modal.com/docs/guide/modal-1-0-migration for more information.",
+    )
 
     def wrapper(f: Union[Callable[[Any], Any], _PartialFunction]) -> _PartialFunction:
         if isinstance(f, _PartialFunction):